+++ platforms/opensolar/auctions.go
@@ -4,33 +4,6 @@ import (
 	"github.com/pkg/errors"
 )
 
-// Contract auctions are specific for public infrastructure and for projects with multiple stakeholders
-// where transparency is important. The main point is to avoid corruption and produce positive competition
-// among providers. When funding solar project using international investor money, transparency and auditability
-// to them is also a crucial aspect since they are often not close to the project, and want to make sure there
-// is a tobust due diligence prior to unlocking the funds.
-
-// Different Auctions or Tenders are designed based on the nature of the project.
-// In general, the criteria for selection is price, technical quality (eg. hardware), engineering model, development time
-// and other perks offered by developers (eg. extra guarantees).
-
-// you need to have a lock in period beyond which contractors can not post what
-// stuff they want. now, how do you choose which contractor wins? Ideally,
-// the school would want the most stuff but you need to vet which contracts are good
-// and not.
-
-// For more trivia, see https://en.wikipedia.org/wiki/Auction
-
-// ContractAuction is an auction struct
-type ContractAuction struct {
-	// TODO: this struct isn't used yet as it needs handlers and stuff, but when
-	// we move off main.go for testing, this must be used in order to make stuff
-	// easier for us.
-	AllContracts    []Project
-	AllContractors  []Entity
-	WinningContract Project
-}
-
 // SelectContractBlind selects the winning bid  based on blind auctio nrules
 // in a blind auction, the bid with the highest price wins
 func SelectContractBlind(arr []Project) (Project, error) {
