+++ platforms/opensolar/developer.go
@@ -1,0 +1,3 @@
+package opensolar
+
+// this should contain the future developer related functions once we define them concretely
