--- platforms/opensolar/entity_developer.go
@@ -1,9 +1,0 @@
-package opensolar
-
-// this should contain the future developer related functions once we define them concretely
-
-// NewDeveloper creates a new developer
-func NewDeveloper(uname string, pwd string, seedpwd string, Name string,
-	Address string, Description string) (Entity, error) {
-	return newEntity(uname, pwd, seedpwd, Name, Address, Description, "developer")
-}
