+++ platforms/opensolar/models.go
@@ -1,4 +1,4 @@
-package munibond
+package opensolar
 
 import (
 	"log"
@@ -12,7 +12,6 @@ import (
 	utils "github.com/Varunram/essentials/utils"
 	consts "github.com/YaleOpenLab/openx/consts"
 	database "github.com/YaleOpenLab/openx/database"
-	models "github.com/YaleOpenLab/openx/models"
 	notif "github.com/YaleOpenLab/openx/notif"
 	oracle "github.com/YaleOpenLab/openx/oracle"
 	"github.com/pkg/errors"
@@ -268,7 +267,54 @@ func MunibondPayback(issuerPath string, recpIndex int, amount float64, recipient
 	return ownershipPct, nil
 }
 
-// SendUSDToPlatform is used to send usd back to the platform
+
+// the models package won't be imported directly in any place but would be imported
+// by all the investment models that exist
+
+// SendUSDToPlatform sends STABLEUSD back to the platform for investment
 func SendUSDToPlatform(invSeed string, invAmount float64, memo string) (string, error) {
-	return models.SendUSDToPlatform(invSeed, invAmount, memo)
+	// send stableusd to the platform (not the issuer) since the issuer will be locked
+	// and we can't use the funds. We also need ot be able to redeem the stablecoin for fiat
+	// so we can't burn them
+	var oldPlatformBalance float64
+	var err error
+	oldPlatformBalance, err = xlm.GetAssetBalance(consts.PlatformPublicKey, consts.StablecoinCode)
+	if err != nil {
+		// platform does not have stablecoin, shouldn't arrive here ideally
+		oldPlatformBalance = 0
+	}
+
+	var txhash string
+	if !consts.Mainnet {
+		_, txhash, err = assets.SendAsset(consts.StablecoinCode, consts.StablecoinPublicKey, consts.PlatformPublicKey, invAmount, invSeed, memo)
+		if err != nil {
+			return txhash, errors.Wrap(err, "sending stableusd to platform failed")
+		}
+	} else {
+		_, txhash, err = assets.SendAsset(consts.AnchorUSDCode, consts.AnchorUSDAddress, consts.PlatformPublicKey, invAmount, invSeed, memo)
+		if err != nil {
+			return txhash, errors.Wrap(err, "sending stableusd to platform failed")
+		}
+	}
+
+	log.Println("Sent STABLEUSD to platform, confirmation: ", txhash)
+	time.Sleep(5 * time.Second) // wait for a block
+
+	var newPlatformBalance float64
+	if !consts.Mainnet {
+		newPlatformBalance, err = xlm.GetAssetBalance(consts.PlatformPublicKey, consts.StablecoinCode)
+		if err != nil {
+			return txhash, errors.Wrap(err, "error while getting asset balance")
+		}
+	} else {
+		newPlatformBalance, err = xlm.GetAssetBalance(consts.PlatformPublicKey, consts.AnchorUSDCode)
+		if err != nil {
+			return txhash, errors.Wrap(err, "error while getting asset balance")
+		}
+	}
+
+	if newPlatformBalance-oldPlatformBalance < invAmount-1 {
+		return txhash, errors.New("Sent amount doesn't match with investment amount")
+	}
+	return txhash, nil
 }
