+++ platforms/opensolar/platform.go
@@ -1,15 +1,6 @@
 package opensolar
 
 import (
-	"crypto/tls"
-	"encoding/json"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"time"
-
-	consts "github.com/YaleOpenLab/openx/consts"
-	notif "github.com/YaleOpenLab/openx/notif"
 	platform "github.com/YaleOpenLab/openx/platforms"
 )
 
@@ -226,7 +217,42 @@ type SustainabilityHelper struct {
 	LCA            string
 }
 
-//easyjson:json
+// Feedback defines a structure that can be used for providing feedback about entities
+type Feedback struct {
+	Content string
+	// the content of the feedback, good / bad
+	// maybe we could have a rating system baked in? a star based rating system?
+	// would be nice, idk
+	From Entity
+	// who gave the feedback?
+	To Entity
+	// regarding whom is this feedback about
+	Date string
+	// time at which this feedback was written
+	Contract []Project
+	// the contract regarding which this feedback is directed at
+}
+
+// Stage is the evolution of the erstwhile static stage integer construction
+type Stage struct {
+	Number          int
+	FriendlyName    string   // the informal name that one can use while referring to the stage (nice for UI as well)
+	Name            string   // this is a more formal name to give to the given stage
+	Activities      []string // the activities that are covered in this particular stage and need to be fulfilled in order to move to the next stage.
+	StateTrigger    []string // trigger state change from n to n+1
+	BreachCondition []string // define breach conditions for a particular stage
+}
+
+// ContractAuction is an auction struct
+type ContractAuction struct {
+	// TODO: this struct isn't used yet as it needs handlers and stuff, but when
+	// we move off main.go for testing, this must be used in order to make stuff
+	// easier for us.
+	AllContracts    []Project
+	AllContractors  []Entity
+	WinningContract Project
+}
+
 type SolarProjectArray []Project
 
 // InitializePlatform imports handlers from the main platform struct that are necessary for starting the platform
@@ -238,68 +264,4 @@ func InitializePlatform() error {
 // is less than 21 XLM, it proceeds to ask the friendbot for more test xlm
 func RefillPlatform(publicKey string) error {
 	return platform.RefillPlatform(publicKey)
-}
-
-const tellerUrl = "https://localhost"
-
-type statusResponse struct {
-	Code   int
-	Status string
-}
-
-// MonitorTeller monitors a teller and checks whether its live. If not, send an email to platform admins
-func MonitorTeller(projIndex int) {
-	// call this function only after a specific order has been accepted by the recipient
-	for {
-		project, err := RetrieveProject(projIndex)
-		if err != nil {
-			log.Println(err)
-			continue
-		}
-
-		tr := &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		}
-		client := &http.Client{Transport: tr}
-
-		req, err := http.NewRequest("GET", tellerUrl+"/ping", nil)
-		if err != nil {
-			log.Println("did not create new GET request", err)
-			notif.SendTellerDownEmail(project.Index, project.RecipientIndex)
-			time.Sleep(consts.TellerPollInterval)
-			continue
-		}
-
-		req.Header.Set("Origin", "localhost")
-		res, err := client.Do(req)
-		if err != nil {
-			log.Println("did not make request", err)
-			notif.SendTellerDownEmail(project.Index, project.RecipientIndex)
-			time.Sleep(consts.TellerPollInterval)
-			continue
-		}
-		data, err := ioutil.ReadAll(res.Body)
-		if err != nil {
-			log.Println("error while reading response body", err)
-			notif.SendTellerDownEmail(project.Index, project.RecipientIndex)
-			time.Sleep(consts.TellerPollInterval)
-			continue
-		}
-
-		var x statusResponse
-		err = json.Unmarshal(data, &x)
-		if err != nil {
-			log.Println("error while unmarshalling data", err)
-			notif.SendTellerDownEmail(project.Index, project.RecipientIndex)
-			time.Sleep(consts.TellerPollInterval)
-			continue
-		}
-
-		if x.Code != 200 || x.Status != "HEALTH OK" {
-			notif.SendTellerDownEmail(project.Index, project.RecipientIndex)
-		}
-
-		res.Body.Close()
-		time.Sleep(consts.TellerPollInterval)
-	}
 }
