+++ platforms/opensolar/platform_test.go
